+++ tests/integration/integration2_test.go
@@ -2,79 +2,18 @@ package integration_test
 
 import (
 	"bytes"
-	"encoding/json"
 	"io"
 	"net/http"
 	"net/http/httptest"
 	"net/url"
 	"strings"
 	"testing"
-	"time"
 
 	"github.com/gin-gonic/gin"
 	"github.com/temirov/llm-proxy/internal/proxy"
-	"go.uber.org/zap"
 )
 
-const (
-	availableModelsBody     = `{"data":[{"id":"gpt-4.1"},{"id":"gpt-5-mini"}]}`
-	webSearchQueryParameter = "web_search"
-)
-
-// roundTripperFunc stubs both models and responses endpoints.
-type roundTripperFunc func(httpRequest *http.Request) (*http.Response, error)
-
-func (roundTripper roundTripperFunc) RoundTrip(httpRequest *http.Request) (*http.Response, error) {
-	return roundTripper(httpRequest)
-}
-
-// makeHTTPClient returns a stub HTTP client capturing payloads and returning canned responses.
-func makeHTTPClient(testingInstance *testing.T, wantWebSearch bool) (*http.Client, *map[string]any) {
-	testingInstance.Helper()
-	var captured map[string]any
-	return &http.Client{
-		Transport: roundTripperFunc(func(httpRequest *http.Request) (*http.Response, error) {
-			switch httpRequest.URL.String() {
-			case proxy.ModelsURL():
-				body := availableModelsBody
-				return &http.Response{StatusCode: http.StatusOK, Body: io.NopCloser(strings.NewReader(body)), Header: make(http.Header)}, nil
-			case proxy.ResponsesURL():
-				if httpRequest.Body != nil {
-					buf, _ := io.ReadAll(httpRequest.Body)
-					_ = json.Unmarshal(buf, &captured)
-				}
-				text := integrationOKBody
-				if wantWebSearch {
-					text = integrationSearchBody
-				}
-				body := `{"output_text":"` + text + `"}`
-				return &http.Response{StatusCode: http.StatusOK, Body: io.NopCloser(strings.NewReader(body)), Header: make(http.Header)}, nil
-			default:
-				testingInstance.Fatalf("unexpected request to %s", httpRequest.URL.String())
-				return nil, nil
-			}
-		}),
-		Timeout: 5 * time.Second,
-	}, &captured
-}
-
-// newLogger constructs a development logger for tests.
-func newLogger(testingInstance *testing.T) *zap.SugaredLogger {
-	testingInstance.Helper()
-	loggerInstance, _ := zap.NewDevelopment()
-	testingInstance.Cleanup(func() { _ = loggerInstance.Sync() })
-	return loggerInstance.Sugar()
-}
-
-// configureProxy sets URLs and the HTTP client for proxy operations.
-func configureProxy(testingInstance *testing.T, client *http.Client) {
-	testingInstance.Helper()
-	proxy.HTTPClient = client
-	proxy.SetModelsURL(mockModelsURL)
-	proxy.SetResponsesURL(mockResponsesURL)
-	testingInstance.Cleanup(proxy.ResetModelsURL)
-	testingInstance.Cleanup(proxy.ResetResponsesURL)
-}
+const webSearchQueryParameter = "web_search"
 
 // TestClientResponseDelivery validates responses with and without web search.
 func TestClientResponseDelivery(testingInstance *testing.T) {
@@ -100,7 +39,7 @@ func TestClientResponseDelivery(testingInstance *testing.T) {
 				QueueSize:     8,
 			}, newLogger(subTest))
 			if buildRouterError != nil {
-				subTest.Fatalf("BuildRouter failed: %v", buildRouterError)
+				subTest.Fatalf(buildRouterFailedFormat, buildRouterError)
 			}
 			server := httptest.NewServer(router)
 			subTest.Cleanup(server.Close)
@@ -114,24 +53,24 @@ func TestClientResponseDelivery(testingInstance *testing.T) {
 			requestURL.RawQuery = queryValues.Encode()
 			httpResponse, requestError := http.Get(requestURL.String())
 			if requestError != nil {
-				subTest.Fatalf("GET failed: %v", requestError)
+				subTest.Fatalf(getFailedFormat, requestError)
 			}
 			defer httpResponse.Body.Close()
 			if httpResponse.StatusCode != http.StatusOK {
-				subTest.Fatalf("status=%d want=%d", httpResponse.StatusCode, http.StatusOK)
+				subTest.Fatalf(statusWantFormat, httpResponse.StatusCode, http.StatusOK)
 			}
 			responseBytes, _ := io.ReadAll(httpResponse.Body)
 			if string(responseBytes) != testCase.expected {
-				subTest.Fatalf("body=%q want=%q", string(responseBytes), testCase.expected)
+				subTest.Fatalf(bodyMismatchFormat, string(responseBytes), testCase.expected)
 			}
 			if testCase.checkTools {
 				tools, ok := (*captured)["tools"].([]any)
 				if !ok || len(tools) == 0 {
-					subTest.Fatalf("tools missing in payload when web_search=1; captured=%v", *captured)
+					subTest.Fatalf(toolsMissingFormat, *captured)
 				}
 				first, _ := tools[0].(map[string]any)
 				if first["type"] != "web_search" {
-					subTest.Fatalf("tool type=%v want=web_search", first["type"])
+					subTest.Fatalf(toolTypeMismatchFormat, first["type"])
 				}
 			}
 		})
@@ -170,7 +109,7 @@ func TestIntegrationConfiguration(testingInstance *testing.T) {
 			if testCase.expectError != "" {
 				_, buildRouterError := proxy.BuildRouter(testCase.config, newLogger(subTest))
 				if buildRouterError == nil || !strings.Contains(buildRouterError.Error(), testCase.expectError) {
-					subTest.Fatalf("expected %s error, got %v", testCase.expectError, buildRouterError)
+					subTest.Fatalf(expectedErrorFormat, testCase.expectError, buildRouterError)
 				}
 				return
 			}
@@ -178,7 +117,7 @@ func TestIntegrationConfiguration(testingInstance *testing.T) {
 			configureProxy(subTest, client)
 			router, buildRouterError := proxy.BuildRouter(testCase.config, newLogger(subTest))
 			if buildRouterError != nil {
-				subTest.Fatalf("BuildRouter failed: %v", buildRouterError)
+				subTest.Fatalf(buildRouterFailedFormat, buildRouterError)
 			}
 			server := httptest.NewServer(router)
 			subTest.Cleanup(server.Close)
@@ -189,13 +128,13 @@ func TestIntegrationConfiguration(testingInstance *testing.T) {
 			requestURL.RawQuery = queryValues.Encode()
 			httpResponse, requestError := http.Get(requestURL.String())
 			if requestError != nil {
-				subTest.Fatalf("GET failed: %v", requestError)
+				subTest.Fatalf(getFailedFormat, requestError)
 			}
 			defer httpResponse.Body.Close()
 			if httpResponse.StatusCode != testCase.expectedStatus {
 				var bodyBuffer bytes.Buffer
 				_, _ = io.Copy(&bodyBuffer, httpResponse.Body)
-				subTest.Fatalf("status=%d want=%d body=%q", httpResponse.StatusCode, testCase.expectedStatus, bodyBuffer.String())
+				subTest.Fatalf(statusWantBodyFormat, httpResponse.StatusCode, testCase.expectedStatus, bodyBuffer.String())
 			}
 		})
 	}
