+++ tests/integration/long_request_test.go
@@ -14,17 +14,8 @@ import (
 )
 
 const (
-	serviceSecretValue = "sekret"
-	openAIKeyValue     = "sk-test"
-	// logLevelDebug is the logging level used in integration tests.
-	logLevelDebug                = "debug"
-	mockModelsURL                = "https://mock.local/v1/models"
-	mockResponsesURL             = "https://mock.local/v1/responses"
 	modelsListBody               = `{"data":[{"id":"gpt-4.1"}]}`
 	expectedResponseBody         = "SLOW_OK"
-	promptQueryParameter         = "prompt"
-	keyQueryParameter            = "key"
-	promptValue                  = "ping"
 	responseDelay                = 31 * time.Second
 	httpClientTimeout            = responseDelay + 5*time.Second
 	requestTimeoutSecondsDefault = 40
@@ -34,15 +25,15 @@ const (
 func makeSlowHTTPClient(testingInstance *testing.T) *http.Client {
 	testingInstance.Helper()
 	return &http.Client{
-		Transport: roundTripperFunc(func(request *http.Request) (*http.Response, error) {
-			switch request.URL.String() {
+		Transport: roundTripperFunc(func(httpRequest *http.Request) (*http.Response, error) {
+			switch httpRequest.URL.String() {
 			case proxy.ModelsURL():
 				return &http.Response{StatusCode: http.StatusOK, Body: io.NopCloser(strings.NewReader(modelsListBody)), Header: make(http.Header)}, nil
 			case proxy.ResponsesURL():
 				time.Sleep(responseDelay)
 				return &http.Response{StatusCode: http.StatusOK, Body: io.NopCloser(strings.NewReader(`{"output_text":"` + expectedResponseBody + `"}`)), Header: make(http.Header)}, nil
 			default:
-				testingInstance.Fatalf("unexpected request to %s", request.URL.String())
+				testingInstance.Fatalf(unexpectedRequestFormat, httpRequest.URL.String())
 				return nil, nil
 			}
 		}),
@@ -59,7 +50,7 @@ func TestIntegrationResponseDeliveredAfterDelay(testingInstance *testing.T) {
 			configureProxy(subTest, makeSlowHTTPClient(subTest))
 			router, buildError := proxy.BuildRouter(proxy.Configuration{ServiceSecret: serviceSecretValue, OpenAIKey: openAIKeyValue, LogLevel: logLevelDebug, WorkerCount: 1, QueueSize: 8, RequestTimeoutSeconds: requestTimeoutSecondsDefault}, newLogger(subTest))
 			if buildError != nil {
-				subTest.Fatalf("BuildRouter failed: %v", buildError)
+				subTest.Fatalf(buildRouterFailedFormat, buildError)
 			}
 			server := httptest.NewServer(router)
 			subTest.Cleanup(server.Close)
@@ -70,15 +61,15 @@ func TestIntegrationResponseDeliveredAfterDelay(testingInstance *testing.T) {
 			requestURL.RawQuery = queryValues.Encode()
 			httpResponse, requestError := http.Get(requestURL.String())
 			if requestError != nil {
-				subTest.Fatalf("GET failed: %v", requestError)
+				subTest.Fatalf(getFailedFormat, requestError)
 			}
 			defer httpResponse.Body.Close()
 			if httpResponse.StatusCode != http.StatusOK {
-				subTest.Fatalf("status=%d want=%d", httpResponse.StatusCode, http.StatusOK)
+				subTest.Fatalf(statusWantFormat, httpResponse.StatusCode, http.StatusOK)
 			}
 			responseBytes, _ := io.ReadAll(httpResponse.Body)
 			if string(responseBytes) != expectedResponseBody {
-				subTest.Fatalf("body=%q want=%q", string(responseBytes), expectedResponseBody)
+				subTest.Fatalf(bodyMismatchFormat, string(responseBytes), expectedResponseBody)
 			}
 		})
 	}
